Course Import and Feed SVGs

Finished GCCISD import system and parsing of retrieved documents into course data.
Connected setup pages
Created system to skip to feed if profile has already been setup
Fixed constraints in activity_feed.xml
Fixed errors with svg icons and reverted them to a less ugly naming scheme.
Replaced old custom add note icon (which did not render correctly) with a new icon from the internet.

+++ app/src/main/java/dev/prognitio/cacao/activities/CourseSetup.java
@@ -48,7 +48,7 @@ public class CourseSetup extends AppCompatActivity {
             String courseName = courseNameInput.getText().toString();
             int courseSemester = -1;
             double courseGPA = -1;
-            double courseGrade = -1;
+            int courseGrade = -1;
             boolean noFormattingErrors = true;
 
 
@@ -65,7 +65,7 @@ public class CourseSetup extends AppCompatActivity {
                 noFormattingErrors = false;
             }
             try {
-                courseGrade = Double.parseDouble(courseGradeInput.getText().toString());
+                courseGrade = Integer.parseInt(courseGradeInput.getText().toString());
             } catch (Exception e) {
                 courseGradeInput.setTextColor(getResources().getColor(R.color.red));
                 noFormattingErrors = false;
@@ -90,7 +90,7 @@ public class CourseSetup extends AppCompatActivity {
             String courseName = courseNameInput.getText().toString();
             int courseSemester = -1;
             double courseGPA = -1;
-            double courseGrade = -1;
+            int courseGrade = -1;
             boolean noFormattingErrors = true;
 
 
@@ -107,7 +107,7 @@ public class CourseSetup extends AppCompatActivity {
                 noFormattingErrors = false;
             }
             try {
-                courseGrade = Double.parseDouble(courseGradeInput.getText().toString());
+                courseGrade = Integer.parseInt(courseGradeInput.getText().toString());
             } catch (Exception e) {
                 courseGradeInput.setTextColor(getResources().getColor(R.color.red));
                 noFormattingErrors = false;
